+++ lib/core/message/file_message.g.dart
@@ -58,7 +58,7 @@ FileMessage _$FileMessageFromJson(Map<String, dynamic> json) => FileMessage(
       isSilent: json['silent'] as bool? ?? false,
       errorCode: json['error_code'] as int?,
       isOperatorMessage: json['is_op_msg'] as bool? ?? false,
-      data: json['data'] as String?,
+      data: BaseMessage._fromJsonToStringData(json['data']),
       ogMetaData: json['og_tag'] == null
           ? null
           : OGMetaData.fromJson(json['og_tag'] as Map<String, dynamic>),