+++ lib/core/message/scheduled_file_message.g.dart
@@ -50,7 +50,7 @@ ScheduledFileMessage _$ScheduledFileMessageFromJson(
           .toList(),
       customType: json['custom_type'] as String?,
       isSilent: json['silent'] as bool? ?? false,
-      data: json['data'] as String?,
+      data: BaseMessage._fromJsonToStringData(json['data']),
       sendingStatus: $enumDecodeNullable(
           _$MessageSendingStatusEnumMap, json['sending_status']),
       isPinnedMessage: json['is_pinned_message'] as bool? ?? false,