+++ lib/core/message/user_message.g.dart
@@ -54,7 +54,7 @@ UserMessage _$UserMessageFromJson(Map<String, dynamic> json) => UserMessage(
       isSilent: json['silent'] as bool? ?? false,
       errorCode: json['error_code'] as int?,
       isOperatorMessage: json['is_op_msg'] as bool? ?? false,
-      data: json['data'] as String?,
+      data: BaseMessage._fromJsonToStringData(json['data']),
       ogMetaData: json['og_tag'] == null
           ? null
           : OGMetaData.fromJson(json['og_tag'] as Map<String, dynamic>),